Initial market node work
